Bump to Meteor 2.6
The package lookback:emails has a dependency to
a really old version of email. Luckily there is a fork on Github which
is compatible with newer versions.
The package raix:push also needed to be removed because of an old
dependency. This packages is only for sending push notifications on
Android and iOS. Since we are not using the Android/iOS build of the app
it should be safe to remove.